Make new core java lib module. Move FSM there
