+++ build.gradle
@@ -1,10 +1,14 @@
 import com.getstream.sdk.chat.Dependencies
+import com.getstream.sdk.chat.DependenciesKt
 
 apply plugin: 'com.pascalwelsch.gitversioner'
 apply plugin: "com.github.ben-manes.versions"
 apply plugin: "org.jlleitschuh.gradle.ktlint"
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
+    ext {
+        kotlin_version = DependenciesKt.KOTLIN_VERSION
+    }
     repositories {
         maven { url "https://plugins.gradle.org/m2/" }
         google()
