+++ stream-chat-core/build.gradle
@@ -1,0 +1,26 @@
+import com.getstream.sdk.chat.Dependencies
+
+plugins {
+    id 'java-library'
+    id 'kotlin'
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_1_8
+    targetCompatibility = JavaVersion.VERSION_1_8
+}
+
+tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
+    kotlinOptions {
+        freeCompilerArgs += [
+                '-progressive',
+                '-Xexplicit-api=strict'
+        ]
+    }
+}
+
+dependencies {
+    implementation Dependencies.kotlinStdLib
+    implementation Dependencies.kotlinReflect
+    implementation Dependencies.coroutinesCore
+}
