+++ stream-chat-core/src/main/java/io/getstream/chat/core/functional/fsm/Event.kt
@@ -1,0 +1,3 @@
+package io.getstream.chat.core.functional.fsm
+
+public interface Event
