+++ stream-chat-core/src/main/java/io/getstream/chat/core/functional/fsm/StateFunction.kt
@@ -1,0 +1,3 @@
+package io.getstream.chat.core.functional.fsm
+
+internal typealias StateFunction<S, E> = FiniteStateMachine<S, E>.(S, E) -> S
